chore: restructure components to have ValidateVoter folder
