+++ src/components/ValidateVoter/EligibleVoterScreen.jsx
