+++ src/components/ValidateVoter/ValidFailPage.jsx
